First pass at splitting API
