+++ api/src/config/index.ts
@@ -1,0 +1,9 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+module.exports = {
+  endpoint: process.env.API_URL,
+  masterKey: process.env.API_KEY,
+  port: process.env.PORT,
+};
